Sanity check automatic renaming of notes
I got a report by Lukas Muehlethaler that he lost several notes while
using my notes.vim plug-in. The problem seems to be related to notes
with Unicode filenames. I've fortunately never had this happen to me and
I can't seem to reproduce his problem but if this really is a bug in my
plug-in it's a serious bug: The plug-in should never lose a single note!

I've now added a sanity check that makes sure the new note exists on
disk before the old note is deleted and if this fails a confirmation
dialog will pop up warning the user of the problem and asking them to
report the problem to me.