Remove notes_fold_ignore_code option, always enabled now!
I have over 800 notes and just tried the :RecentNotes command for the
first time in a while. It was really slow, took more than a minute..
I tried the :ShowTaggedNotes command, also very slow.

It turns out that the "improvement" to text folding which ignores # inside
code blocks was written in a very naive way, making it very slow on large
buffers (over 500 lines). I've now fixed this, making it fast enough on
large buffers to always enable the improvement and remove the option
altogether. I like it when I can remove options :-).