Improve compatibility with Txtfmt plug-in
As I documented in a previous commit, the txtfmt plug-in can be used
together with the notes.vim plug-in by setting the notes file type
to "notes.txtfmt". This had the unfortunate side effect of making it
impossible to save notes, because the notes plug-in checks in several
places that &filetype == 'notes'. I've now changed the code to recognize
the case where the &filetype option contains multiple file type names.