Support case sensitive keyword searches (issue #34)
This change breaks backwards compatibility with the previous keyword
index format, however the Python script will recreate the index in
the new format and everything should be fine.

Also this change enlarges the file size of the index, but I have ideas
on how to speed up the search-notes.py script (to be continued :-).