Disable special key mappings inside {{{ snippets }}}
This is my first serious attempt to resolve issue #79 and issue #83.
I was originally worried that this might have serious impact on typing
speed because the <expr> mappings now have to search back through the
current buffer to determine their context.

In practice the slowness isn't that bad and it *is* really nice to type
regular quotes inside code blocks. No doubt this will become a performance
problem for someone editing 10.000 line notes, but we'll cross that bridge
when the get there ;-)